feature: add the dotmatrix theme
The dotmatrix theme is basically the look of "powerline", but done right on the user experience side.
The main idea is to highlight what's important, instead of displaying too much things because fancy colors everywhere is fun.

To achieve that, the dotmatrix theme has been very carefully designed:
- it uses three four modules: host/path, machine state, environment, version control.
- it uses a consistent visual grammar (using 3 separators and 3 colors).
- it follows good practices for color blindness.

The current code use a similar approach to the one used for the powerline theme, but needs a refactoring.